requirements.txt
 pyTelegramBotAPI>=4.6.0 
 peewee>=3.14.0
 PyYAML>=6.0